+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
-## Head (0.7.4)
+## Head (0.7.5)
+* switch from wp_nav_menu to wp_list_items
 * coding guideline update: csscomb, convert tabs to spaces (based on codeguide.co)
 * add the most important attributes to hamburger at nav.scss
 * added marker.click preset to googlemaps template